fix(deploy/check): modify check items level
