+++ pkg/deploy/operation/check/check_docker_version_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package docker
+package check
 
 import (
 	"fmt"
