+++ pkg/deploy/operation/check/check_system_distribution.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package system
+package check
 
 import (
 	"fmt"
@@ -36,7 +36,7 @@ const (
 
 type CheckDistributionOperation struct {
 	operation.BaseOperation
-	operation.CheckOperations
+	CheckOperations
 }
 
 func (ckops *CheckDistributionOperation) getScript() string {
